scripts/requirements.txt
python-frontmatter==1.0.0
beautifulsoup4==4.12.1
Markdown==3.4.3